requirements.txt
celery==5.0.5
channels==3.0.3
channels-redis==3.2.0
daphne==3.0.1
Django==3.2.5
django-active-link==0.1.7
django-celery-beat==2.2.0
django-cors-middleware==1.5.0
django-filter==2.4.0
django-redis==4.12.1
django-rest-swagger==2.2.0
django-rest-auth[with_social]==0.9.5
djangorestframework==3.12.2
djangorestframework-filters==0.10.2
gevent==21.1.2
gunicorn==20.0.4
psutil==5.8.0
psycopg2-binary==2.8.6
pre-commit==2.12.1
pylint==2.8.2
pylint-django==2.4.4
python-decouple==3.4
redis==3.5.3
requests==2.25.1
